[main] Initially move custom logic to own file
As a first step to refactor haulage to be more modular and
customizeable, refactor all of the custom logic specific to the UW
deployment out to its own file.
